updated code js to ts
