+++ .gitignore
@@ -5,4 +5,4 @@ node_modules
 
 /src/generated/prisma
 
-.dist
+dist/
