+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+      "module": "esnext",
+      "esModuleInterop": true,
+      "target": "es2017",
+      "moduleResolution": "node",
+      "sourceMap": true,
+      "outDir": "dist",
+      "types": ["node"]
+    },
+    "lib": ["es2015"]
+}
